--- Harsha/Angular/Crud/models/user.ts
@@ -1,5 +1,0 @@
-export interface User {
-    id: number;
-    firstName: string;
-    lastName: string;
-}
