--- Harsha/Angular/app/app.component.css
@@ -1,4 +1,0 @@
-div h2{
-    text-align: center;
-    color: red;
-}
