+++ Harsha/Angular/app/app.component.html
@@ -1,7 +1,94 @@
-<div>
-  <h2>
-    Hi, This is the {{title}}
-  </h2>
+<div class="container">
+  <div class="row">
+    <div class="col-md-12 mt-4">
+      <h1 class="text-center">User List</h1>
+      <div class="filter-wrap">
+       <input [(ngModel)]="searchTerm" placeholder="Filter Users Based on First Name" class="filter-input">
+      </div>
+      <table class="table table-bordered table-striped table-hover">
+        <thead>
+          <tr>
+            <th class="text-center">ID</th>
+            <th class="text-center">First Name</th>
+            <th class="text-center">Last Name</th>
+            <th class="text-center" >Address</th>
+            <th class="text-center">State</th>
+            <th class="text-center">Email</th>
+            <th class="text-center">&nbsp;</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr *ngFor="let user of filterUser">
+            <td>{{ user.id }}</td>
+            <td>{{ user.firstName }}</td>
+            <td>{{ user.lastName }}</td>
+            <td>{{ user.address }}</td>
+            <td>{{user.state}}</td>
+            <td>{{user.email}}</td>
+
+            <td>
+              <button type="button" class="btn btn-info btn-sm mr-2" (click)="showEditUserForm(user)">Edit</button>
+              <button type="button" class="btn btn-danger btn-sm" (click)="removeUser(user)">Delete</button>
+            </td>
+          </tr>
+        </tbody>
+      </table>
+      <button class="btn btn-success" (click)="showAddUserForm()">Add New User</button>
+      <div class="mt-4">
+        <form *ngIf="userForm">
+          <h2>Add New User</h2>
+          <div class=" form-group">
+            <form class="form" #registerForm="ngForm">        
+          <p>First Name: <input type="text" name="firstName"  [(ngModel)]="newUser.firstName"  #firstName="ngModel" required />
+             <span class="text-danger"
+              *ngIf="(firstName.touched || registerForm.submitted) && firstName.errors?.required">First Name is required
+              </span>
+          </p>
+          <p>Last Name: <input type="text" name="lastName" #lastName="ngModel" required  [(ngModel)]="newUser.lastName" />
+            <span class="text-danger"
+            *ngIf="(lastName.touched || registerForm.submitted) && lastName.errors?.required">Last Name is required
+            </span>
+          </p>
+          <p>Address  : <textarea type="text" name="address"  #address="ngModel" required [(ngModel)]="newUser.address" ></textarea>
+            <span class="text-danger"
+            *ngIf="(address.touched || registerForm.submitted) && address.errors?.required">Address is required
+            </span>
+          </p>
+          <p>State  :<select style="width: 20%;" name="state" [(ngModel)]="newUser.state">
+            <option >Choose State</option>
+            <option *ngFor="let state of states" >{{state.name}}</option> 
+          </select> </p>
+          <p>Email Id  :<input type="text" name="email" #email="ngModel" required  [(ngModel)]="newUser.email" />
+            <span class="text-danger"
+             *ngIf="(email.touched || registerForm.submitted) && email.errors?.required">Email is required
+            </span></p>
+          </form>
+          <p>
+            <button type="button" class="btn btn-primary mr-2" (click)="saveUser(newUser)">Submit</button>
+            <button type="button" class="btn btn-warning" (click)="cancelNewUser()">Cancel</button>
+          </p>
+        </div>
+        </form>
+
+        <form *ngIf="editUserForm">
+          <h2>Edit User Details</h2>
+          <p>First Name: <input type="text" name="firstName" [(ngModel)]="editedUser.firstName" /></p>
+          <p>Last Name: <input type="text" name="lastName" [(ngModel)]="editedUser.lastName" /></p>
+          <p>Address  : <textarea type="text" name="address" [(ngModel)]="editedUser.address" ></textarea></p>
+          <p>State  :<select style="width: 20%;" name="state" [(ngModel)]="editedUser.state">
+            <option >Choose State</option>
+            <option *ngFor="let state of states"  [value]="state.name">{{state.name}}</option>  
+          </select></p>
+          <p>Email Id : <select  style="width: 20%;" name="email" [(ngModel)]="editedUser.email" >
+            <option >Choose Email</option>
+            <option *ngFor="let user of users" [value]="user.email">{{user.email}}</option>  
+          </select></p>
+          <p>
+            <button type="button" class="btn btn-primary mr-2" (click)="updateUser()">Submit</button>
+            <button type="button" class="btn btn-warning" (click)="cancelEdits()">Cancel</button>
+          </p>
+        </form>
+      </div>
+    </div>
+  </div>
 </div>
-<app-artists></app-artists>
-<app-artistnames></app-artistnames>
